backend/requirements.txt
asgiref==3.2.10
attrs==20.2.0
cached-property==1.5.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1
Django==3.1.1
django-cors-headers==3.5.0
django-filter==2.3.0
django-rest-knox==4.1.0
djangorestframework==3.11.1
enum-compat==0.0.3
Faker==4.1.3
idna==2.10
iniconfig==1.0.1
Markdown==3.2.2
more-itertools==8.5.0
packaging==20.4
pluggy==0.13.1
psycopg2==2.8.6
py==1.9.0
pycparser==2.20
pyparsing==2.4.7
pytest==6.0.2
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
urllib3==1.25.10
